requirements.txt
opendataloader-pdf
fastapi
uvicorn
jira>=3.8.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
requests>=2.31.0
rich>=13.7.0
numpy>=1.26.0
openai>=1.30.0,<2.0.0
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0
python-multipart
loguru>=0.7.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
mcp>=1.0.0
# PageIndex dependencies
litellm>=1.82.0
pymupdf>=1.26.0
PyPDF2>=3.0.1
